+++ Changelog.md
@@ -4,6 +4,7 @@
 * added a noisy simulator backend to simulate errors of real quantum hardware
 * use a single state atom in simulator
 * added perfect power factor to math
+* added a more robust circuit optimization pipeline
 * generalized quantum phase estimation and integrated in quantum-period-finding
 * phase estimation handles the number of measurements, reduced the number of circuit creations
 * fixed edge cases for quantum phase estimation